test: validate variable in action inside foreach container (for builder ui)
