+++ keep-ui/entities/workflows/lib/__tests__/validation.test.ts
@@ -1,161 +1,6 @@
-import {
-  validateMustacheVariableName,
-  validateAllMustacheVariablesInString,
-  validateStepPure,
-  validateGlobalPure,
-} from "../validate-definition";
+import { validateStepPure, validateGlobalPure } from "../validate-definition";
 import { Provider } from "@/shared/api/providers";
 import { Definition, V2Step } from "../../model/types";
-
-describe("validateMustacheVariableName", () => {
-  const mockDefinition: Definition = {
-    sequence: [
-      {
-        id: "step1",
-        name: "First Step",
-        componentType: "task",
-        type: "step-test",
-        properties: {
-          actionParams: [],
-          stepParams: [],
-        },
-      },
-      {
-        id: "step2",
-        name: "Second Step",
-        componentType: "task",
-        type: "action-test",
-        properties: {
-          actionParams: [],
-          stepParams: [],
-        },
-      },
-    ],
-    properties: {
-      id: "test-workflow",
-      name: "Test Workflow",
-      description: "Test Description",
-      disabled: false,
-      isLocked: false,
-      consts: {},
-    },
-  };
-  const mockSecrets = {};
-
-  it("should validate alert variables", () => {
-    const result = validateMustacheVariableName(
-      "{{ alert.name }}",
-      mockDefinition.sequence[0],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toBeNull();
-  });
-
-  it("should validate incident variables", () => {
-    const result = validateMustacheVariableName(
-      "{{ incident.title }}",
-      mockDefinition.sequence[0],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toBeNull();
-  });
-
-  it("should validate step results access", () => {
-    const result = validateMustacheVariableName(
-      "{{ steps.First Step.results }}",
-      mockDefinition.sequence[1],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toBeNull();
-  });
-
-  it("should prevent accessing current step results", () => {
-    const result = validateMustacheVariableName(
-      "{{ steps.First Step.results }}",
-      mockDefinition.sequence[0],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toBe(
-      "Variable: '{{ steps.First Step.results }}' - You can't access the results of the current step."
-    );
-  });
-
-  it("should prevent accessing future step results", () => {
-    const result = validateMustacheVariableName(
-      "{{ steps.Second Step.results }}",
-      mockDefinition.sequence[0],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toBe(
-      "Variable: '{{ steps.Second Step.results }}' - You can't access the results of a step that appears after the current step."
-    );
-  });
-
-  it("should prevent accessing action results from a step", () => {
-    const result = validateMustacheVariableName(
-      "{{ steps.Second Step.results }}",
-      mockDefinition.sequence[0],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toBe(
-      "Variable: '{{ steps.Second Step.results }}' - You can't access the results of a step that appears after the current step."
-    );
-  });
-});
-
-describe("validateAllMustacheVariablesInString", () => {
-  const mockDefinition: Definition = {
-    sequence: [
-      {
-        id: "step1",
-        name: "First Step",
-        componentType: "task",
-        type: "step-test",
-        properties: {
-          actionParams: [],
-          stepParams: [],
-        },
-      },
-    ],
-    properties: {
-      id: "test-workflow",
-      name: "Test Workflow",
-      description: "Test Description",
-      disabled: false,
-      isLocked: false,
-      consts: {},
-    },
-  };
-  const mockSecrets = {};
-
-  it("should validate multiple variables in a string", () => {
-    const result = validateAllMustacheVariablesInString(
-      "Alert: {{ alert.name }} with severity {{ alert.severity }}",
-      mockDefinition.sequence[0],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toEqual([]);
-  });
-
-  it("should detect invalid variables in a string", () => {
-    const result = validateAllMustacheVariablesInString(
-      "Invalid: {{ invalid.var }} and {{ steps.Future Step.results }}",
-      mockDefinition.sequence[0],
-      mockDefinition,
-      mockSecrets
-    );
-    expect(result).toContain(
-      "Variable: '{{ steps.Future Step.results }}' - a 'Future Step' step that doesn't exist."
-    );
-  });
-});
 
 describe("validateStepPure", () => {
   const mockProviders: Provider[] = [
